fix(trpc): cast magicLink.analytics median to int
Mirrors the existing count()::int treatment — pg can return
double precision as a string depending on driver parse config,
which would silently break the declared `number | null` return
type. Dwell is integer-ms anyway, so rounding is lossless.

+++ apps/web/src/server/trpc/routers/magic-link.ts
@@ -111,23 +111,25 @@ export const magicLinkRouter = router({
     // with zero views. Notes:
     // - COUNT(views.id) (not COUNT(*)) so unmatched LEFT JOIN rows
     //   yield 0, not 1.
-    // - COUNT returns bigint, which the pg driver hands back as a
-    //   string; ::int cast keeps the wire type a JS number (safe — a
-    //   single producer's view tally won't exceed 2^31).
+    // - Both COUNT and PERCENTILE_CONT are cast to ::int for the same
+    //   reason: COUNT returns bigint and PERCENTILE_CONT returns
+    //   double precision, either of which the pg driver may hand back
+    //   as a string; the cast keeps the wire type a JS number. Safe
+    //   here because a single producer's view tally won't exceed 2^31,
+    //   and dwell_ms is stored as integer milliseconds (sub-ms median
+    //   precision is noise — rounding is lossless).
     // - PERCENTILE_CONT(0.5) is the only ordered-set aggregate that
     //   gives true median (AVG would be skewed by a few long tails);
     //   over an empty set it returns NULL, which we propagate as-is so
     //   the UI can distinguish "no data" from "0ms".
     // - GROUP BY just magicLinks.id — PG infers the rest of the
     //   selected magic_links columns via PK functional dependency.
-    //   (None are selected here, but the same rule lets us add them
-    //   later without touching GROUP BY.)
     return ctx.db
       .select({
         id: magicLinks.id,
         viewCount: sql<number>`count(${magicLinkViews.id})::int`,
         lastViewedAt: sql<Date | null>`max(${magicLinkViews.viewedAt})`,
-        medianDwellMs: sql<number | null>`percentile_cont(0.5) within group (order by ${magicLinkViews.dwellMs})`,
+        medianDwellMs: sql<number | null>`percentile_cont(0.5) within group (order by ${magicLinkViews.dwellMs})::int`,
       })
       .from(magicLinks)
       .leftJoin(magicLinkViews, eq(magicLinkViews.magicLinkId, magicLinks.id))
